fix all test
the all test on department entity are fixed
