+++ src/test/java/com/example/stream/spring/courses/reactive/example/controller/InstructorControllerTest.java
@@ -2,7 +2,6 @@ package com.example.stream.spring.courses.reactive.example.controller;
 
 import com.example.stream.spring.courses.reactive.example.model.request.InstructorRequestDto;
 import com.example.stream.spring.courses.reactive.example.model.response.InstructorResponseDto;
-import com.example.stream.spring.courses.reactive.example.model.response.StudentResponseDto;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -32,23 +31,6 @@ class InstructorControllerTest {
         assertNotNull(webTestClient);
     }
 
-    /**
-     * Tests the retrieval of students by teacher ID.
-     */
-    @Test
-    void testGetStudentsByTeacher() {
-        Long teacherId = 1L;
-        StudentResponseDto student1 = new StudentResponseDto("1L", "Alice", "Nel paese", "alice@example.com");
-        StudentResponseDto student2 = new StudentResponseDto("2L", "Bob", "Patigno", "bob@example.com");
-
-        webTestClient.get()
-                .uri("/teacher/{teacherId}/students", teacherId)
-                .exchange()
-                .expectStatus().isOk()
-                .expectBodyList(StudentResponseDto.class)
-                .hasSize(2)
-                .contains(student1, student2);
-    }
 
     @Test
     void find_all_instructor() {
